# api/src/main/resources/public/api.yaml
openapi: 3.0.1
info:
  title: Car Hire API
  version: 0.1.0
  contact:
    name: Loredana Costea
  description: Car Hire API
servers:
  - url: http://localhost:{port}
    description: local environment
    variables:
      port:
        default: '8080'
tags:
  - name: user
    description: user endpoint
  - name: fleet
    description: fleet endpoint
  - name: rental
    description: rental endpoint
security:
  - oAuth:
      - write
      - read
paths:
  /user:
    get:
      tags:
        - user
      summary: get users
      description: returns a list of all the users, without any filtering
      operationId: getUsers
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - user
      summary: add a new user
      description: adds a new user and returns it
      operationId: addUser
      requestBody:
        description: The data of the user to insert. The id and status will be generated automatically.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
            example:
              email: loredana.costea@gmail.com
              password: 1234
              username: loredna
              firstName: Costea
              lastName: Loredana
              country: Romania
              licenseNumber: 999888
      responses:
        '201':
          description: the created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      tags:
        - user
      summary: update an existing user
      description: updates the first name, last name, country, license number, status of an existing user and returns it
      operationId: updateUser
      requestBody:
        description: The data of the user to update. The id can't be changed.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
            example:
              lastName: Loredana Iuliana
              id: 1
      responses:
        '200':
          description: the updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /user/{id}:
    get:
      tags:
        - user
      summary: get a user
      description: returns the details for a user, based on id
      operationId: getUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - user
      summary: removes a user
      description: deletes a user identified by id
      operationId: removeUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /fleet:
    get:
      tags:
        - car
      summary: get cars
      description: returns a list of all the cars, without any filtering
      operationId: getCars
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - car
      summary: add a new car
      description: adds a new car and returns it
      operationId: addCar
      requestBody:
        description: The data of the car to insert. The id and status will be generated automatically.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
            example:
              make: BMW
              model: X6
              year: 2019
              mileage: 40000
              fuel: Diesel
              sizeClass: Compact
      responses:
        '200':
          description: the created car
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      tags:
        - car
      summary: update an existing car
      description: updates the make, model, year, mileage, fuel, status of an existing car and returns it
      operationId: updateCar
      requestBody:
        description: The data of the car to update. The id can't be changed.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
            example:
              model: I3
              id: 1
      responses:
        '200':
          description: the created car
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /fleet/{id}:
    get:
      tags:
        - car
      summary: get a car
      description: returns the details for a car, based on id
      operationId: getCar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - car
      summary: removes a car
      description: deletes a car identified by id
      operationId: removeCar
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /rental:
    get:
      tags:
        - rental
      summary: get rentals
      description: returns a list of all the reservations, without any filtering
      operationId: getRentals
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RentalDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - rental
      summary: add a new rental
      description: adds a new reservation and returns it
      operationId: addRental
      requestBody:
        description: The data of the rental to insert. The id and status will be generated automatically.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalDTO'
            example:
              userId: 1
              carId: 100
              startDateTime: 2021-01-30T08:30:00Z
              endDateTime: 2021-04-30T08:30:00Z
      responses:
        '200':
          description: the created rental
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      tags:
        - rental
      summary: update an existing rental
      description: updates the startDateTime, endDateTime, status of an existing rezervation and returns it
      operationId: updateRental
      requestBody:
        description: The data of the rental to update. The id can't be changed.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalDTO'
            example:
              status: CANCELLED
              id: 1
      responses:
        '200':
          description: the created rental
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /rental/{id}:
    get:
      tags:
        - rental
      summary: get a rental
      description: returns the details for a reservation, based on id
      operationId: getRental
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - rental
      summary: removes a rental
      description: deletes a rental identified by id
      operationId: removeRental
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
    Unauthorized:
      description: unauthorized
    BadRequest:
      description: bad request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ValidationDTO'
    ServerError:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
  schemas:
    ValidationDTO:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        description:
          type: string
    ErrorDTO:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: string
          description: id of the user
        email:
          type: string
          description: email of the user
        password:
          type: string
          description: password of the user
        username:
          type: string
          description: username of the user
        firstName:
          type: string
          description: first name of the user
        lastName:
          type: string
          description: last name of the user
        country:
          $ref: '#/components/schemas/CountryDTO'
        licenseNumber:
          type: string
          description: license number of the user
        status:
          type: string
          enum:
            - ACTIVE
            - DELETED
          description: status of the user
        title:
          type: string
          description: title of the user
    CountryDTO:
      type: object
      properties:
        id:
          type: string
          description: country id
        name:
          type: string
          example: Romania
          description: the name of the country
        isoCode:
          type: string
          example: ROU
          description: the iso code of the country
    CarDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: id of the car
        make:
          type: string
          example: BMW, FORD, AUDI
          description: make of the car
        model:
          type: string
          example: X5, 7 Series, A3, I8
          description: model of the car
        year:
          type: integer
          description: year of the car
        mileage:
          type: integer
          format: int64
          description: mileage of the car
        fuel:
          type: string
          example: Diesel
          description: fuel type of the car
        sizeClass:
          type: string
          example: Minicompact, Compact, Mid-Size, Large
          description: size class of the car
        status:
          type: string
          enum:
            - ACTIVE
            - DELETED
          description: status of the car
    RentalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: id of the rental
        userId:
          type: integer
          format: int32
          description: id of the user
        carId:
          type: integer
          format: int32
          description: id of the car
        startDateTime:
          type: string
          format: date-time
          description: the start date and time on which the reservation was created
        endDateTime:
          type: string
          format: date-time
          description: the end date and time on which the reservation was cancelled
        status:
          type: string
          enum:
            - ACTIVE
            - CANCELLED
          description: status of the car
  securitySchemes:
    oAuth:
      type: oauth2
      description: This API uses OAuth2 with the password flow
      flows:
        password:
          tokenUrl: /oauth/token
          scopes:
            read: get information from the API
            write: use the API to make changes to the supporting data model